schnorr: Remove unused threshold code.
This removes the completely unused threshold code from the schnorr
package.  This code should not be used as it only implements naive key
combining which is easily exploitable without extreme care.  There is no
reason to keep this code around.
