+++ dcrec/secp256k1/schnorr/ecdsa.go
@@ -153,6 +153,17 @@ func schnorrSign(msg []byte, ps []byte, k []byte,
 	return &Signature{Rpx, sBig}, nil
 }
 
+// nonceRFC6979 is a local instantiation of deterministic nonce generation
+// by the standards of RFC6979.
+func nonceRFC6979(privkey []byte, hash []byte, extra []byte, version []byte, extraIterations uint32) []byte {
+	pkD := new(big.Int).SetBytes(privkey)
+	defer pkD.SetInt64(0)
+	bigK := secp256k1.NonceRFC6979(pkD, hash, extra, version, extraIterations)
+	defer bigK.SetInt64(0)
+	k := bigIntToEncodedBytes(bigK)
+	return k[:]
+}
+
 // Sign is the exported version of sign. It uses RFC6979 and Blake256 to
 // produce a Schnorr signature.
 func Sign(priv *secp256k1.PrivateKey, hash []byte) (r, s *big.Int, err error) {
